Emulate textinput through key presses
Quest version cannot fetch intended unicode character from pressed key,
because required functions are not exposed to JNI. As a workaround this
commit implements basic and most used character combinations based on
US keyboard layout. It is not very likely that Quest will ever implement
selection of keyboatd layout, so Lua is better place for it anyway.