Remove recovery mode and revise errhand
The recovery mode and errorpane are removed together with the mechanism
to restart lovr and catch errors on startup. This complication was
introduced so that mistakes in editor code could be fixed in safe env.

With self-modifying aspect dropped the error handling becomes simpler -
just stop executing the user code and show the stack trace.

Jump to line functionality is gone for now. Too complicated and user
should already be on the line that caused problems in runtime.